// src/app/generated/model/studentsReviewDataDto.ts
/**
 * swagger
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReviewModuleWithValueDto } from './reviewModuleWithValueDto';


export interface StudentsReviewDataDto { 
    reviewId?: number;
    reviewer?: string | null;
    topicName?: string | null;
    englishTopicName?: string | null;
    modules?: Array<ReviewModuleWithValueDto> | null;
}